requirements.txt
fastapi[all]
uvicorn[standard]
pymongo
ultralytics
opencv-python
numpy
pillow
python-multipart
